Moved subquery tests to their own dedicated test package
