+++ go/test/endtoend/vtgate/misc_test.go
@@ -423,20 +423,6 @@ func TestShowVariables(t *testing.T) {
 	require.True(t, found, "Expected a row for version in show query")
 }
 
-func TestSubQueryOnTopOfSubQuery(t *testing.T) {
-	defer cluster.PanicHandler(t)
-	ctx := context.Background()
-	conn, err := mysql.Connect(ctx, &vtParams)
-	require.NoError(t, err)
-	defer conn.Close()
-	defer utils.Exec(t, conn, `delete from t1`)
-
-	utils.Exec(t, conn, `insert into t1(id1, id2) values (1, 1), (2, 2), (3, 3), (4, 4), (5, 5)`)
-	utils.Exec(t, conn, `insert into t2(id3, id4) values (1, 3), (2, 4)`)
-
-	utils.AssertMatches(t, conn, "select id1 from t1 where id1 not in (select id3 from t2) and id2 in (select id4 from t2) order by id1", `[[INT64(3)] [INT64(4)]]`)
-}
-
 func TestShowVGtid(t *testing.T) {
 	conn, err := mysql.Connect(context.Background(), &vtParams)
 	require.NoError(t, err)
@@ -483,17 +469,6 @@ func TestShowGtid(t *testing.T) {
 		require.True(t, exists, "gtid not cached for row: %v", row)
 		require.NotEqual(t, gtid, row[1].ToString())
 	}
-}
-
-func TestQueryAndSubQWithLimit(t *testing.T) {
-	conn, err := mysql.Connect(context.Background(), &vtParams)
-	require.NoError(t, err)
-	defer conn.Close()
-
-	defer utils.Exec(t, conn, `delete from t1`)
-	utils.Exec(t, conn, "insert into t1(id1, id2) values(0,0),(1,1),(2,2),(3, 3), (4, 4), (5, 5), (6, 6), (7, 7), (8, 8),(9,9)")
-	result := utils.Exec(t, conn, `select id1, id2 from t1 where id1 >= ( select id1 from t1 order by id1 asc limit 1) limit 100`)
-	assert.Equal(t, 10, len(result.Rows))
 }
 
 func TestDeleteAlias(t *testing.T) {
@@ -580,19 +555,6 @@ func TestRenameFieldsOnOLAP(t *testing.T) {
 	_ = utils.Exec(t, conn, "use mysql")
 	qr = utils.Exec(t, conn, "select @@workload")
 	assert.Equal(t, `[[VARBINARY("OLAP")]]`, fmt.Sprintf("%v", qr.Rows))
-}
-
-func TestSubqueriesHasValues(t *testing.T) {
-	defer cluster.PanicHandler(t)
-	ctx := context.Background()
-	conn, err := mysql.Connect(ctx, &vtParams)
-	require.NoError(t, err)
-	defer conn.Close()
-
-	defer utils.Exec(t, conn, `delete from t1`)
-	utils.Exec(t, conn, "insert into t1(id1, id2) values (0,1),(1,2),(2,3),(3,4),(4,5),(5,6)")
-	utils.AssertMatches(t, conn, `SELECT id2 FROM t1 WHERE id1 IN (SELECT id1 FROM t1 WHERE id1 > 10)`, `[]`)
-	utils.AssertMatches(t, conn, `SELECT id2 FROM t1 WHERE id1 NOT IN (SELECT id1 FROM t1 WHERE id1 > 10) ORDER BY id2`, `[[INT64(1)] [INT64(2)] [INT64(3)] [INT64(4)] [INT64(5)] [INT64(6)]]`)
 }
 
 func TestSelectEqualUniqueOuterJoinRightPredicate(t *testing.T) {
