+++ go/test/endtoend/vtgate/queries/subquery/main_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package union
+package subquery
 
 import (
 	"flag"
